WorkOrderBuildHut extends WorkOrderBuildDecoration instead of WorkOrderBuildDecoration extends WorkOrderBuild
