+++ src/main/java/com/minecolonies/coremod/entity/ai/citizen/builder/ConstructionTapeHelper.java
@@ -3,8 +3,8 @@ package com.minecolonies.coremod.entity.ai.citizen.builder;
 import com.minecolonies.coremod.blocks.AbstractBlockHut;
 import com.minecolonies.coremod.blocks.ModBlocks;
 import com.minecolonies.coremod.colony.buildings.AbstractBuilding;
-import com.minecolonies.coremod.colony.workorders.WorkOrderBuild;
 import com.minecolonies.coremod.colony.workorders.WorkOrderBuildDecoration;
+import com.minecolonies.coremod.colony.workorders.WorkOrderBuildHut;
 import com.minecolonies.coremod.configuration.Configurations;
 import com.minecolonies.coremod.util.BlockUtils;
 import com.minecolonies.coremod.util.StructureWrapper;
@@ -46,7 +46,7 @@ public final class ConstructionTapeHelper
      */
     public static void placeConstructionTape(@NotNull AbstractBuilding building, @NotNull World world)
     {
-        placeConstructionTape(new WorkOrderBuild(building, 1), world);
+        placeConstructionTape(new WorkOrderBuildHut(building, 1), world);
     }
 
     /**
@@ -56,7 +56,7 @@ public final class ConstructionTapeHelper
      * @param world     the world.
      */
 
-    public static void placeConstructionTape(@NotNull WorkOrderBuild workOrder, @NotNull World world)
+    public static void placeConstructionTape(@NotNull WorkOrderBuildDecoration workOrder, @NotNull World world)
     {
         if (Configurations.builderPlaceConstructionTape)
         {
@@ -197,7 +197,7 @@ public final class ConstructionTapeHelper
      */
     public static void removeConstructionTape(@NotNull AbstractBuilding building, @NotNull World world)
     {
-        removeConstructionTape(new WorkOrderBuild(building, 1), world);
+        removeConstructionTape(new WorkOrderBuildHut(building, 1), world);
     }
 
     /**
@@ -207,12 +207,12 @@ public final class ConstructionTapeHelper
      * @param world     the world.
      */
 
-    public static void removeConstructionTape(@NotNull WorkOrderBuild workOrder, @NotNull World world)
+    public static void removeConstructionTape(@NotNull WorkOrderBuildDecoration workOrder, @NotNull World world)
     {
         final StructureWrapper wrapper = new StructureWrapper(world, workOrder.getStructureName());
         final BlockPos pos = workOrder.getBuildingLocation();
         int tempRotation = 0;
-        if (workOrder.getRotation() == 0 && !(workOrder instanceof WorkOrderBuildDecoration))
+        if (workOrder.getRotation() == 0 && workOrder instanceof WorkOrderBuildHut)
         {
             final IBlockState blockState = world.getBlockState(pos);
             if (blockState.getBlock() instanceof AbstractBlockHut)
