+++ src/main/java/com/minecolonies/coremod/entity/ai/citizen/builder/EntityAIStructureBuilder.java
@@ -7,8 +7,8 @@ import com.minecolonies.coremod.colony.buildings.AbstractBuildingWorker;
 import com.minecolonies.coremod.colony.buildings.BuildingBuilder;
 import com.minecolonies.coremod.colony.buildings.utils.BuildingBuilderResource;
 import com.minecolonies.coremod.colony.jobs.JobBuilder;
-import com.minecolonies.coremod.colony.workorders.WorkOrderBuild;
 import com.minecolonies.coremod.colony.workorders.WorkOrderBuildDecoration;
+import com.minecolonies.coremod.colony.workorders.WorkOrderBuildHut;
 import com.minecolonies.coremod.configuration.Configurations;
 import com.minecolonies.coremod.entity.ai.basic.AbstractEntityAIStructure;
 import com.minecolonies.coremod.entity.ai.util.AIState;
@@ -99,7 +99,7 @@ public class EntityAIStructureBuilder extends AbstractEntityAIStructure<JobBuild
      */
     private void loadStructure()
     {
-        WorkOrderBuild workOrder = null;
+        WorkOrderBuildDecoration workOrder = null;
         workOrder = job.getWorkOrder();
 
         if (workOrder == null)
@@ -108,7 +108,7 @@ public class EntityAIStructureBuilder extends AbstractEntityAIStructure<JobBuild
         }
 
         final BlockPos pos = workOrder.getBuildingLocation();
-        if (!(workOrder instanceof WorkOrderBuildDecoration) && worker.getColony().getBuilding(pos) == null)
+        if (workOrder instanceof WorkOrderBuildHut && worker.getColony().getBuilding(pos) == null)
         {
             Log.getLogger().warn("AbstractBuilding does not exist - removing build request");
             worker.getColony().getWorkManager().removeWorkOrder(workOrder);
@@ -116,7 +116,7 @@ public class EntityAIStructureBuilder extends AbstractEntityAIStructure<JobBuild
         }
 
         int tempRotation = 0;
-        if (workOrder.getRotation() == 0 && !(workOrder instanceof WorkOrderBuildDecoration))
+        if (workOrder.getRotation() == 0 && workOrder instanceof WorkOrderBuildHut)
         {
             final IBlockState blockState = world.getBlockState(pos);
             if (blockState.getBlock() instanceof AbstractBlockHut)
@@ -153,9 +153,9 @@ public class EntityAIStructureBuilder extends AbstractEntityAIStructure<JobBuild
             return true;
         }
 
-        final WorkOrderBuild wo = job.getWorkOrder();
+        final WorkOrderBuildDecoration wo = job.getWorkOrder();
 
-        if (job.getColony().getBuilding(wo.getBuildingLocation()) == null && !(wo instanceof WorkOrderBuildDecoration))
+        if (job.getColony().getBuilding(wo.getBuildingLocation()) == null && wo instanceof WorkOrderBuildHut)
         {
             job.complete();
             return true;
@@ -315,7 +315,7 @@ public class EntityAIStructureBuilder extends AbstractEntityAIStructure<JobBuild
         {
             workFrom = null;
             loadStructure();
-            final WorkOrderBuild wo = job.getWorkOrder();
+            final WorkOrderBuildDecoration wo = job.getWorkOrder();
             if (wo == null)
             {
                 Log.getLogger().error(
@@ -459,7 +459,8 @@ public class EntityAIStructureBuilder extends AbstractEntityAIStructure<JobBuild
         final String structureName = job.getStructure().getName();
         worker.sendLocalizedChat("entity.builder.messageBuildComplete", structureName);
 
-        final WorkOrderBuild wo = job.getWorkOrder();
+        final WorkOrderBuildDecoration wo = job.getWorkOrder();
+        final WorkOrderBuildHut woh = (wo instanceof WorkOrderBuildHut)?(WorkOrderBuildHut) wo : null;
         if (wo == null)
         {
             Log.getLogger().error(String.format("Builder (%d:%d) ERROR - Finished, but missing work order(%d)",
@@ -469,11 +470,11 @@ public class EntityAIStructureBuilder extends AbstractEntityAIStructure<JobBuild
         }
         else
         {
-            if (wo instanceof WorkOrderBuildDecoration && structureName.contains(WAYPOINT_STRING))
+            if (woh == null && structureName.contains(WAYPOINT_STRING))
             {
                 worker.getColony().addWayPoint(wo.getBuildingLocation(), world.getBlockState(wo.getBuildingLocation()));
             }
-            else
+            else if (woh != null)
             {
                 final AbstractBuilding building = job.getColony().getBuilding(wo.getBuildingLocation());
                 if (building == null)
@@ -481,11 +482,11 @@ public class EntityAIStructureBuilder extends AbstractEntityAIStructure<JobBuild
                     Log.getLogger().error(String.format("Builder (%d:%d) ERROR - Finished, but missing building(%s)",
                             worker.getColony().getID(),
                             worker.getCitizenData().getId(),
-                            wo.getBuildingLocation()));
+                            woh.getBuildingLocation()));
                 }
                 else
                 {
-                    building.setBuildingLevel(wo.getUpgradeLevel());
+                    building.setBuildingLevel(woh.getUpgradeLevel());
                 }
             }
             job.complete();
